fwctl: FWCTL_INFO to return basic information about the device
Userspace will need to know some details about the fwctl interface being
used to locate the correct userspace code to communicate with the
kernel. Provide a simple device_type enum indicating what the kernel
driver is.

Allow the device to provide a device specific info struct that contains
any additional information that the driver may need to provide to
userspace.

Signed-off-by: Jason Gunthorpe <jgg@nvidia.com>
Link: https://lore.kernel.org/netdev/2-v2-940e479ceba9+3821-fwctl_jgg@nvidia.com/t/#mcd777f4efbabd43ab8674ee7ee14ed3be16e418f
Signed-off-by: Tushar Dave <tdave@nvidia.com>