taint: Add TAINT_FWCTL
Requesting a fwctl scope of access that includes mutating device debug
data will cause the kernel to be tainted. Changing the device operation
through things in the debug scope may cause the device to malfunction in
undefined ways. This should be reflected in the TAINT flags to help any
debuggers understand that something has been done.

Signed-off-by: Jason Gunthorpe <jgg@nvidia.com>
Link: https://lore.kernel.org/netdev/2-v2-940e479ceba9+3821-fwctl_jgg@nvidia.com/t/#m6218cdf62843b9f40c9a53a7401eacc349272259
Signed-off-by: Tushar Dave <tdave@nvidia.com>